删除广告图片添加的功能，这个已经在广告详细页面里面有了 后台订单添加商品的时候可以给予促销了

+++ admin/add_images/remove.php
@@ -18,8 +18,6 @@
 ?>
 <?php require_once('../../Connections/localhost.php'); ?>
 <?php
-$doc_url="ad.html#delete_image";
-$support_email_question="删除广告图片";
 $could_delete=1;
 $colname_news = "-1";
 
@@ -35,24 +33,22 @@ $ad_images = mysql_query($query_ad_images, $localhost) or die(mysql_error());
 $row_ad_images=mysql_fetch_assoc($ad_images);
 $totalRows_ad_images = mysql_num_rows($ad_images);
 if($totalRows_ad_images==0){
- 	$remove_succeed_url="/admin/ad/detail.php?recordID=".$row_ad_images['ad_id'];
-	header("Location: " . $remove_succeed_url );
+	$could_delete=0;
 } 
 
- 
 if($could_delete==1){
 	
-	$if_result=@unlink($_SERVER['DOCUMENT_ROOT'].$row_ad_images['image_path']);
-	 
-	$update_catalog = sprintf("delete from `ad_images` where id = %s", $colname_ad_images);
-	$update_catalog_query = mysql_query($update_catalog, $localhost) or die(mysql_error());
-	if(!$update_catalog_query){
-		$could_delete=0;
-	}else{
-		$remove_succeed_url="/admin/ad/detail.php?recordID=".$row_ad_images['ad_id'];
-		header("Location: " . $remove_succeed_url );
+	$if_result=unlink($_SERVER['DOCUMENT_ROOT'].$row_ad_images['image_path']);
+	if($if_result){
+		$update_catalog = sprintf("delete from `ad_images` where id = %s", $colname_ad_images);
+		$update_catalog_query = mysql_query($update_catalog, $localhost);
+		if(!$update_catalog_query){
+			$could_delete=0;
+		}else{
+ 			$remove_succeed_url="/admin/ad/detail.php?recordID=".$row_ad_images['ad_id'];
+			header("Location: " . $remove_succeed_url );
+		}
 	}
-	 
 	
 }
 
@@ -66,13 +62,21 @@ if($could_delete==1){
 </head>
 
 <body>
-<?php if($could_delete==0){ ?>
 <div class="phpshop123_infobox">
-	  <p>由于以下原因，您不能删除这张图片，请及时修正，或是联系123phpshop.com的技术支持人员！</p>
-	  <p>1. 图片不存在，请检查参数之后再试。</p>
-	  <p>2. 系统错误，无法删除，请稍后再试。 </p>
-	  <p>您也可以<a href="index.php">点击这里返回</a>。</p>
+  <?php if($could_delete==0){ ?>
+  <p>由于以下原因，您不能删除这张图片：</p>
+  <p>1. 图片不存在，请检查参数之后再试。</p>
+  <p>2. 系统错误，无法删除，请示稍后再试。 </p>
+  <p>您也可以<a href="index.php">点击这里返回</a>。</p>
+  <?php } ?>
 </div>
-<?php } ?>
+
 </body>
 </html>
+<?php
+mysql_free_result($news);
+
+mysql_free_result($ad_images);
+?>
+
+
